requirements.txt
streamlit
snowflake-connector-python
pandas
sqlparse
